backend/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Search Engine
opensearch-py==2.3.1
opensearch-dsl==2.1.0

# AI/LLM Integrations
openai==1.28.1
anthropic==0.28.0  # Claude API
langchain==0.1.14
instructor==1.0.0  # Structured LLM outputs
sentence-transformers==2.2.2  # Embeddings
torch==2.2.2
huggingface-hub==0.25.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Logging & Observability
structlog==23.2.0
python-json-logger==2.0.7
pyyaml==6.0.1

opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Async & Concurrency
aiohttp==3.9.1
asyncpg==0.29.0

# Utilities
requests==2.31.0
tenacity==8.2.3  # Retry logic
python-dateutil==2.8.2
